# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The cloud-native API gateway for the Lexos AI document processing engine.",
        "title": "Lexos API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/glean/ask": {
            "get": {
                "description": "Submits a query against a previously indexed document and streams the AI's answer token-by-token via Server-Sent Events (SSE).",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Gleaner"
                ],
                "summary": "Ask a question against an indexed document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The task_id of the previously indexed document",
                        "name": "document_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The user's question",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE Stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/glean/index": {
            "post": {
                "description": "Streams a document to MinIO, chunks it, embeds it via FastEmbed, and builds a FAISS index in the background.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gleaner"
                ],
                "summary": "Upload document for vector indexing",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Document to index (.txt, .pdf, .docx)",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Document queued for indexing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing document file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the current status of the Lexos Gateway.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Check API health",
                "responses": {
                    "200": {
                        "description": "API is healthy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/summarize": {
            "post": {
                "description": "Uploads a text payload or a document (.txt, .pdf, .docx) to MinIO and queues it for Map-Reduce summarization by Qwen3.",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Distiller"
                ],
                "summary": "Submit a document for AI summarization",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Document to summarize (if using multipart)",
                        "name": "document",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Summary style (bullet_points, short_paragraph, executive)",
                        "name": "style",
                        "in": "formData"
                    },
                    {
                        "description": "JSON payload (if using application/json)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.SummarizeRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Task queued successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "description": "Fetches the current processing state and result URLs for a given task ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Get the status of an async task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task state details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transcribe": {
            "post": {
                "description": "Uploads an audio file to MinIO and queues it for transcription using Faster-Whisper.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scriber"
                ],
                "summary": "Submit audio for transcription",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file to transcribe",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Task queued successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing audio file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.SummarizeRequest": {
            "type": "object",
            "properties": {
                "document_text": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                }
            }
        }
    }
}